% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_arrays.R
\name{add_dimnames}
\alias{add_dimnames}
\title{Add dimension names (to arrays).}
\usage{
add_dimnames(
  x,
  dnames = c("row", "col", "tab"),
  prefix = c("r", "c", "t"),
  sep = ""
)
}
\arguments{
\item{x}{An array to which dimension names are to be added.}

\item{dnames}{A character vector of dimension names. 
If \code{length(dnames == 1)}, a numeric digit is added to 
signal the dimension number. 
If \code{length(dnames)} exceeds the number of dimensions, 
it is truncated. 
If the number of dimensions exceeds \code{length(dnames)}, 
\code{dnames} are created from \code{LETTERS}. 
Default: \code{dnames = c("row", "col", "tab")}.}

\item{prefix}{A character vector of prefixes (whose
length must match the number of dimensions of \code{x}). 
Default: \code{prefix = c("r", "c", "t")} (for row, column, and table).}

\item{sep}{A separator (as character) between prefix and numeric indices. 
Default: \code{sep = ""}.}
}
\value{
Input \code{x} with dimension names.
}
\description{
\code{add_dimnames} adds (or re-assigns) dimension names to arrays 
(or data frames or atomic vectors, but not other lists).
}
\details{
When \code{x} is an atomic vector, 
\code{add_dimnames} returns a named vector. 

See \code{\link{provideDimnames}} (from \strong{base} R) 
for a related function.
}
\examples{
a <- array(1:24, dim = c(3, 4, 2))
add_dimnames(a)
add_dimnames(a, dnames = "dim", sep = "_")
add_dimnames(a, dnames = c("X", "Y", "Z", "XYZ"), prefix = c("x", "y", "z"), sep = "_")

# More dimensions:
b <- array(1:2^5, dim = rep(2, 5))
add_dimnames(b)
add_dimnames(b, dnames = "D")
add_dimnames(b, dnames = "dim", prefix = c("r", "c", "u", "v", "w"), sep = "_")

# For data frames:
c <- data.frame(let = letters[1:4], num = 5:8)
is.list(c)
is.data.frame(c)
add_dimnames(c, prefix = c("r", "c"))
add_dimnames(c)

# For atomic vectors:
d <- 1:4
add_dimnames(d, prefix = "nr")  # adding named elements
add_dimnames(d, prefix = letters[1:4], sep = "_")

# For scalars:
add_dimnames(NA)

# For lists (that are NOT data frames):
l <- list(1:2, letters[1:3])
add_dimnames(l)  # return NA

}
\seealso{
\code{\link{provideDimnames}} is a more basic version.

Other array functions: 
\code{\link{expand_freq_table}()},
\code{\link{flatten_array}()}
}
\concept{array functions}
