% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fun.R
\name{base2dec}
\alias{base2dec}
\title{Convert a base N numeric string (to a decimal numeric string).}
\usage{
base2dec(seq, base = 2)
}
\arguments{
\item{seq}{A (required) sequence or string of numeric symbols 
(i.e., digits 0-9, as defined in specified base).}

\item{base}{The base of the symbols in \code{seq}. 
Default: \code{base = 2} (binary).}
}
\value{
A decimal number.
}
\description{
\code{base2dec} converts a sequence of numeric symbols 
from base N notation to decimal (i.e., base 10) notation.
}
\examples{
# (a) string inputs:
base2dec("0")
base2dec("1")
base2dec("1010")  # seq as string
# (b) numeric vectors as inputs:
base2dec(c(1, 0, 1, 0))
base2dec(c(1, 0, 1, 0), base = 3)

}
\seealso{
\code{\link{as.roman}} converts integers into Roman numerals.
}
\concept{numeric functions}
