% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_arrays.R
\name{ctable}
\alias{ctable}
\title{Create a contingency table from data and description (as a table).}
\usage{
ctable(
  data,
  dim = length(data),
  dimnames = NULL,
  by_row = FALSE,
  as_df = FALSE
)
}
\arguments{
\item{data}{The data to be used (required; 
frequency counts, as a \code{\link{vector}} of positive integers).}

\item{dim}{The dimensions to be used 
(as a numeric \code{\link{vector}}). 
Default: \code{dim = length(data)}.}

\item{dimnames}{The dimension names to be used 
(as a \code{\link{list}}). 
Default: \code{dimnames = NULL}, using 
\code{\link{add_dimnames}} per default.}
}
\value{
A table.
}
\description{
\code{ctable} yields a contingency table (as a \code{table}) 
in analogy to creating an array by \code{\link{array}}.
}
\details{
\code{ctable} internally passes its arguments to \code{\link{array}} 
before coercing the resulting \code{array} into a \code{table} or \code{data.frame}. 

The order of elements in its arguments \code{data}, \code{dim}, 
and \code{dimnames} must correspond to the defaults of \code{\link{array}}. 
For instance, \code{data} are entered in a by-column fashion 
and the dimensions in \code{dimnames} from left/y, top/x, to group/table/z, etc.
}
\examples{
# Prepare dimnames (as lists): 
l_24 <- list(rows = c("A", "B"), cols = letters[1:4])
l432 <- list(Ys = letters[1:4], Xs = LETTERS[1:3], group = c("I", "II"))

# Standard use:
ctable(1:8,  dim = c(2,4), dimnames = l_24)
ctable(1:24, dim = c(4,3,2), dimnames = l432)

# by row:
ctable(1:8,  dim = c(2,4), dimnames = l_24, by_row = TRUE)
ctable(1:24, dim = c(4,3,2), dimnames = l432, by_row = TRUE)

# Return as df:
ctable(1:8, dim = c(2,4), dimnames = l_24, as_df = TRUE)
ctable(1:24, dim = c(4,3,2), dimnames = l432, as_df = TRUE)

# No dimnames:
ctable(1:8, dim = c(2,4))
ctable(1:8, dim = c(2,2,2))

}
\seealso{
\code{\link{subtable}} for extracting a subset of a table;  
\code{\link{sublist}} for extracting subsets of a list;  
\code{\link{table}} and \code{\link{array}} for the underlying data structures.

Other array functions: 
\code{\link{add_dimnames}()},
\code{\link{expand_freq_table}()},
\code{\link{flatten_array}()},
\code{\link{subtable}()}
}
\concept{array functions}
