% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_arrays.R
\name{expand_freq_table}
\alias{expand_freq_table}
\title{Expand a contingency table (as array/table or data frame) into a raw cases (as data frame).}
\usage{
expand_freq_table(x, freq_var = "Freq", row_name_repair = TRUE)
}
\arguments{
\item{x}{An contingency table (as \code{\link{table}} or \code{\link{data.frame}}).}

\item{freq_var}{The name of the frequency count variable in 
a data frame \code{x} (aka. \code{responseName} in \code{\link{as.data.frame}}). 
Default: \code{freq_var = "Freq"},  
based on default of \code{as.data.frame(x, responseName = "Freq")} for arrays/tables.}

\item{row_name_repair}{Boolean: Should rows be repaired (i.e., enumerated)? 
Default: \code{row_name_repair = TRUE}.}
}
\value{
A data frame (of raw cases).
}
\description{
\code{expand_freq_table} turns a contingency table 
(given as an array/table or as a data frame with a frequency count variable 
\code{freq_var} that denotes the number of cases for each factor combination) 
into a data frame of raw cases.
}
\details{
\code{expand_freq_table} assumes that \code{x} is a \code{\link{table}} 
or a \code{\link{data.frame}} with a frequency count variable \code{freq_var}. 

If \code{x} is a \code{\link{table}}, \code{expand_freq_table} first uses 
\code{\link{as.data.frame}} (with \code{responseName = freq_var}) 
to turn \code{x} into a contingency table (as a data frame with 
a frequency variable, named \code{"Freq"} by default). 

The function allows turning a contingency table 
(i.e., a table that cross-classifies frequency counts) ---  
in the form of a \code{\link{table}} or a \code{\link{data.frame}} 
with a frequency count variable \code{freq_var} --- 
into a corresponding \code{\link{data.frame}} of raw cases. 
The number of cases or observations (rows) in the resulting data frame 
corresponds to \code{sum(x)} (for tables) or 
\code{sum(x$freq_var)} (for data frames).
}
\examples{
# (a) from raw data (vectors):
ans <- sample(c("yes", "no", "maybe"), 100, replace = TRUE)
eat <- sample(c("fish", "meat", "veggie"), 100, replace = TRUE)
df_1 <- data.frame(ans, eat)  # data frame from vectors
df_2 <- expand_freq_table(data.frame(table(ans, eat))) # table > contingency table > df
all.equal(table(df_1), table(df_2))

# (b) from a table:
df <- expand_freq_table(UCBAdmissions)  # array/table > contingency table > df
tb <- table(df)                                # df > array/table 
all.equal(UCBAdmissions, tb)

# Trivial case:
expand_freq_table(data.frame(x = "a", Freq = 2))

# Full circle (4D array > contingency table > data frame > 4D array): 
df <- expand_freq_table(Titanic)
tb <- table(df)
all.equal(Titanic, tb)

}
\seealso{
\code{\link{table}} and \code{\link{xtabs}} for turning data frames into contingency tables;  
\code{\link{as.data.frame}} for turning an array/table into a contingency table (as df).

Other array functions: 
\code{\link{add_dimnames}()},
\code{\link{flatten_array}()},
\code{\link{subtable}()}
}
\concept{array functions}
