% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fun.R
\name{base2dec}
\alias{base2dec}
\title{Convert a string of numeric digits from some base into a decimal number.}
\usage{
base2dec(x, base = 2)
}
\arguments{
\item{x}{A (required) sequence or string of numeric symbols 
(i.e., digits 0-9, as defined in the specified base).}

\item{base}{The base of the symbols in \code{seq}. 
Default: \code{base = 2} (binary).}
}
\value{
A decimal number.
}
\description{
\code{base2dec} converts a sequence of numeric symbols 
from base N notation to decimal (i.e., base 10) notation.
}
\details{
\code{base2dec} is the complement of \code{\link{dec2base}}.
}
\examples{
# (a) string inputs:
base2dec("11")
base2dec("11", base = 5)
base2dec("1010")
base2dec("0101")
# (b) numeric vectors as inputs:
base2dec(c(1, 0, 1, 0))
base2dec(c(1, 0, 1, 0), base = 3)

# Note: 
base2dec(dec2base(0120, base = 3), base = 3)
dec2base(base2dec(0210, base = 3), base = 3)

}
\seealso{
\code{\link{dec2base}} converts decimal numbers into another base;  
\code{\link{as.roman}} converts integers into Roman numerals.

Other numeric functions: 
\code{\link{dec2base}()}
}
\concept{numeric functions}
