% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fun.R
\name{dec2base}
\alias{dec2base}
\title{Convert a decimal number into a string of numeric digits in some base.}
\usage{
dec2base(x, base = 2)
}
\arguments{
\item{x}{A (required) integer in decimal notation 
(i.e., containing only digits 0-9).}

\item{base}{The base of the symbols in \code{seq}. 
Default: \code{base = 2} (binary).}
}
\value{
A number (in base notation).
}
\description{
\code{dec2base} converts a decimal (i.e., base 10) number into 
a sequence of numeric symbols (digits) in some other base.
}
\details{
\code{dec2base} is the complement of \code{\link{base2dec}}.
}
\examples{
dec2base(8)
dec2base(8, base = 3)
dec2base(8, base = 7)
dec2base(8, base = 10)

# Note: 
base2dec(dec2base(0120, base = 3), base = 3)
dec2base(base2dec(0210, base = 3), base = 3)

}
\seealso{
\code{\link{base2dec}} converts numbers from some base into decimal numbers;  
\code{\link{as.roman}} converts integers into Roman numerals.

Other numeric functions: 
\code{\link{base2dec}()}
}
\concept{numeric functions}
