% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_arrays.R
\name{expand_data_frame}
\alias{expand_data_frame}
\title{Expand an array/contingency table into a data frame.}
\usage{
expand_data_frame(x, freq_var = "Freq", fix_row_names = TRUE)
}
\arguments{
\item{x}{An array/table or contingency table (as data frame).}

\item{freq_var}{The name of the frequency count variable in 
a data frame \code{x}. 
Default: \code{freq_var = "Freq"} 
(i.e., as in \code{data.frame(x)} for arrays/tables).}

\item{fix_row_names}{Boolean: Should rows be named consecutively? 
Default: \code{fix_row_names = TRUE}.}
}
\value{
A data frame.
}
\description{
\code{expand_data_frame} turns an array/table or a contingency table 
(with a frequency count variable denoting the number of corresponding cases) 
into a data frame of raw cases.
}
\details{
\code{expand_data_frame} assumes that \code{x} is an array/table 
or a data frame with a frequency count variable \code{freq_var}. 

If \code{x} is an array/table, \code{expand_data_frame} first uses 
\code{\link{data.frame}} to turn \code{x} into a contingency table 
(with a frequency variable named \code{"Freq"} by default). 

The function allows turning data stored as an array/table 
or a contingency table (with a frequency count variable \code{freq_var}) 
into a corresponding data frame of raw cases. 
The number of cases (rows) in the resulting data frame 
corresponds to \code{sum(x)} (for tables) or 
\code{sum(x$freq_var)} (for data frames).
}
\examples{
# (a) from raw data (vectors):
ans <- sample(c("yes", "no", "maybe"), 100, replace = TRUE)
eat <- sample(c("fish", "meat", "veggie"), 100, replace = TRUE)
df_1 <- data.frame(ans, eat)  # data frame from vectors
df_2 <- expand_data_frame(data.frame(table(ans, eat))) # from table > contingency table > data frame
all.equal(table(df_1), table(df_2))

# (b) from array/table:
df <- expand_data_frame(UCBAdmissions)  # array/table > contingency table > df
tb <- table(df)                         # df > array/table 
all.equal(UCBAdmissions, tb)

# Trivial case:
expand_data_frame(data.frame(x = "a", Freq = 2))

# Full circle (4D array > contingency table > data frame > 4D array): 
df <- expand_data_frame(Titanic)
tb <- table(df)
all.equal(Titanic, tb)

}
\seealso{
\code{\link{table}} for the inverse function;  
\code{\link{data.frame}} for turning an array/table into a contingency table.

Other array functions: 
\code{\link{add_dimnames}()},
\code{\link{flatten_array}()}
}
\concept{array functions}
