% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_arrays.R
\name{sub_table}
\alias{sub_table}
\title{Get a subset (or sub-table) of a table (as table).}
\source{
Based on the \code{subtable} function by Norman Matloff, 
The Art of R Programming (2011, pp. 131-134).
}
\usage{
sub_table(t, sub_dims)
}
\arguments{
\item{t}{The original table (as \code{\link{table}}).}

\item{sub_dims:}{A list specifying the desired subtable extraction. 
Each element is named after a dimension of \code{t}, 
and the value of that component is a vector of the names or numbers of the desired levels.}
}
\value{
A table.
}
\description{
\code{sub_table} yields a subset of a table 
by filtering or extracting a subset of table's 
dimensions and levels.
}
\details{
\code{sub_table} provides a filter or slice function for tables, 
by specifying a positive subset (i.e., levels to be included).
}
\examples{

}
\seealso{
\code{\link{table}} and \code{\link{array}} for the underlying data structures.

Other array functions: 
\code{\link{add_dimnames}()},
\code{\link{expand_freq_table}()},
\code{\link{flatten_array}()}
}
\concept{array functions}
